Querying the MongoDB works.
Next commit: clean up the leftover code.
Explanation: All find() queries can be achivede with aggregate queries. MongoDB shell syntax is not directly compatible with the Java MongoDB driver. In Java, all the aggregate operators (json objects) have to be parsed to a Document object. They are stored in a List that is given to the aggregate() function.